POS-73: fix variable types from float32 to float64

+++ src/pkg/report/population_growth_leaders/domain/domain.go
@@ -13,5 +13,5 @@ type GrowthLeadersReport struct {
 	OKTMO               int
 	YearFrom            int
 	YearTo              int
-	PopulationGrowthPct float32
+	PopulationGrowthPct float64
 }
